pyproject.toml
[project]
name = "resizer"
version = "0.0.0"
description = "(Batch) Image resizer cli command based on Pillow"
readme = "README.rst"
keywords = ["image", "processing", "cli"]
license = {file = "LICENSE"}
authors = [
    {name = "Grzegorz Śliwiński", email = "fizyk+pypi@fizyk.dev"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "click",
    "pillow",
]
requires-python = ">= 3.11"

[project.scripts]
resizer = "resizer.cli:resizer"

[project.urls]
"Source" = "https://github.com/fizyk/resizer"
"Bug Tracker" = "https://github.com/fizyk/resizer/issues"
"Changelog" = "https://github.com/fizyk/resizer/blob/v0.0.0/CHANGES.rst"

[build-system]
requires = ["setuptools >= 61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = true


[tool.setuptools.packages.find]
include = ["resizer*"]
exclude = ["tests*"]
namespaces = false

[tool.pytest.ini_options]
xfail_strict=true
addopts = "--showlocals --verbose --cov"
testpaths = "tests"

[tool.black]
line-length = 120
target-version = ['py311']
include = '.*\.pyi?$'

[tool.ruff]
# Decrease the maximum line length to 79 characters.
line-length = 120
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
  "D",   # pydocstyle
]

[tool.mypy]
exclude = [
    "build/*",
]

[tool.towncrier]
directory = "newsfragments"
single_file=true
filename="CHANGES.rst"
issue_format="`#{issue} <https://https://github.com/fizyk/resizer/issues/{issue}>`_"

[tool.towncrier.fragment.feature]
name = "Features"
showcontent = true

[tool.towncrier.fragment.bugfix]
name = "Bugfixes"
showcontent = true

[tool.towncrier.fragment.break]
name = "Breaking changes"
showcontent = true

[tool.towncrier.fragment.misc]
name = "Miscellaneus"
showcontent = true

[tool.tbump]
# Uncomment this if your project is hosted on GitHub:
# github_url = "https://github.com/ClearcodeHQ/pytest-dynamodb/"

[tool.tbump.version]
current = "0.0.0"

# Example of a semver regexp.
# Make sure this matches current_version before
# using tbump
regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (\-
    (?P<extra>.+)
  )?
  '''

[tool.tbump.git]
message_template = "Release {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.field]]
# the name of the field
name = "extra"
# the default value to use, if there is no match
default = ""


# For each file to patch, add a [[file]] config
# section containing the path of the file, relative to the
# tbump.toml location.
[[tool.tbump.file]]
src = "resizer/__init__.py"

[[tool.tbump.file]]
src = "pyproject.toml"
search = 'version = "{current_version}"'

[[tool.tbump.file]]
src = "pyproject.toml"
search = '"Changelog" = "https://github.com/fizyk/resizer/blob/v{current_version}/CHANGES.rst"'

# You can specify a list of commands to
# run after the files have been patched
# and before the git commit is made

[[tool.tbump.before_commit]]
name = "Build changelog"
cmd = "pipenv run towncrier build --version {new_version} --yes"

# Or run some commands after the git tag and the branch
# have been pushed:
#  [[tool.tbump.after_push]]
#  name = "publish"
#  cmd = "./publish.sh"